+++ lt/dsp/convolution/FFTConvolver.cpp
@@ -43,14 +43,14 @@ FFTConvolver::FFTConvolver(std::size_t signalSize, std::size_t patchSize)
 {
 }
 
-auto FFTConvolver::convolute(double const* signal, double const* patch, double* output) const -> void
+auto FFTConvolver::convolute(float const* signal, float const* patch, float* output) const -> void
 {
-    std::fill(signalScratch_.get(), std::next(signalScratch_.get(), totalSize_), 0.0);
-    std::fill(patchScratch_.get(), std::next(patchScratch_.get(), totalSize_), 0.0);
-    std::fill(tmp_.get(), std::next(tmp_.get(), totalSize_), 0.0);
-    std::fill(signalScratchOut_.get(), std::next(signalScratchOut_.get(), totalSize_), Complex<double> {});
-    std::fill(patchScratchOut_.get(), std::next(patchScratchOut_.get(), totalSize_), Complex<double> {});
-    std::fill(tmpOut_.get(), std::next(tmpOut_.get(), totalSize_), 0.0);
+    std::fill(signalScratch_.get(), std::next(signalScratch_.get(), totalSize_), 0.0f);
+    std::fill(patchScratch_.get(), std::next(patchScratch_.get(), totalSize_), 0.0f);
+    std::fill(tmp_.get(), std::next(tmp_.get(), totalSize_), 0.0f);
+    std::fill(signalScratchOut_.get(), std::next(signalScratchOut_.get(), totalSize_), Complex<float> {});
+    std::fill(patchScratchOut_.get(), std::next(patchScratchOut_.get(), totalSize_), Complex<float> {});
+    std::fill(tmpOut_.get(), std::next(tmpOut_.get(), totalSize_), 0.0f);
 
     for (auto i = std::size_t { 0 }; i < totalSize_; i++) {
         if (i < signalSize_) {
