+++ lt/dsp/convolution/FFTConvolver.hpp
@@ -4,11 +4,11 @@
 #include "lt/dsp/fft/FFT.hpp"
 
 struct FFTConvolver {
-    using value_type = double;
+    using value_type = float;
 
     FFTConvolver(std::size_t signalSize, std::size_t patchSize);
 
-    auto convolute(double const* signal, double const* patch, double* output) const -> void;
+    auto convolute(float const* signal, float const* patch, float* output) const -> void;
 
 private:
     std::size_t signalSize_;
@@ -18,12 +18,12 @@ private:
     std::unique_ptr<RealFFT> forwardFFT_;
     std::unique_ptr<RealFFT> backwardFFT_;
 
-    std::unique_ptr<double[]> signalScratch_ { std::make_unique<double[]>(totalSize_) };
-    std::unique_ptr<Complex<double>[]> signalScratchOut_ { std::make_unique<Complex<double>[]>(totalSize_) };
+    std::unique_ptr<float[]> signalScratch_ { std::make_unique<float[]>(totalSize_) };
+    std::unique_ptr<Complex<float>[]> signalScratchOut_ { std::make_unique<Complex<float>[]>(totalSize_) };
 
-    std::unique_ptr<double[]> patchScratch_ { std::make_unique<double[]>(totalSize_) };
-    std::unique_ptr<Complex<double>[]> patchScratchOut_ { std::make_unique<Complex<double>[]>(totalSize_) };
+    std::unique_ptr<float[]> patchScratch_ { std::make_unique<float[]>(totalSize_) };
+    std::unique_ptr<Complex<float>[]> patchScratchOut_ { std::make_unique<Complex<float>[]>(totalSize_) };
 
-    std::unique_ptr<Complex<double>[]> tmp_ { std::make_unique<Complex<double>[]>(totalSize_) };
-    std::unique_ptr<double[]> tmpOut_ { std::make_unique<double[]>(totalSize_) };
+    std::unique_ptr<Complex<float>[]> tmp_ { std::make_unique<Complex<float>[]>(totalSize_) };
+    std::unique_ptr<float[]> tmpOut_ { std::make_unique<float[]>(totalSize_) };
 };
