+++ lt/dsp/wavelets/WaveletPacketTransform.hpp
@@ -20,29 +20,29 @@ struct WaveletPacketTransform {
     int MaxIter {}; // Maximum Iterations J <= MaxIter
     std::string ext; // Type of Extension used - "per" or "sym"
     std::string entropy;
-    double eparam {};
+    float eparam {};
 
     int N {}; //
     int nodes {};
     int length[102] {};
-    double* output {};
-    double* costvalues {};
-    double* basisvector {};
+    float* output {};
+    float* costvalues {};
+    float* basisvector {};
     int* nodeindex {};
     int* numnodeslevel {};
     int* coeflength {};
-    std::unique_ptr<double[]> params;
+    std::unique_ptr<float[]> params;
 
 private:
     Wavelet* wave_ { nullptr };
     int signalLength_ {}; // Length of the original signal.
 };
 
-auto dwt(WaveletPacketTransform& wt, double const* inp) -> void;
-auto idwt(WaveletPacketTransform& wt, double* dwtop) -> void;
+auto dwt(WaveletPacketTransform& wt, float const* inp) -> void;
+auto idwt(WaveletPacketTransform& wt, float* dwtop) -> void;
 
 auto setDWPTExtension(WaveletPacketTransform& wt, char const* extension) -> void;
-auto setDWPTEntropy(WaveletPacketTransform& wt, char const* entropy, double eparam) -> void;
+auto setDWPTEntropy(WaveletPacketTransform& wt, char const* entropy, float eparam) -> void;
 auto getDWPTNodelength(WaveletPacketTransform& wt, int x) -> int;
 
 auto summary(WaveletPacketTransform const& wt) -> void;
